backend/requirements.txt
Flask>=3.0,<4
Flask-CORS>=4.0,<5
Flask-SQLAlchemy>=3.1,<4
SQLAlchemy>=2.0.32,<2.1
Flask-Marshmallow>=1.2,<2
marshmallow>=3.21,<4
python-dotenv>=1.0,<2
Flask-Talisman>=1.1,<2
passlib[bcrypt]>=1.7,<2
itsdangerous>=2.2,<3